+++ AmazonChimeSDK/AmazonChimeSDK/session/MeetingSessionConfiguration.swift
@@ -10,7 +10,8 @@ import Foundation
 
 /// `MeetingSessionConfiguration` contains the information necessary to start a session.
 /// Constructs a MeetingSessionConfiguration with a chime:`CreateMeetingResponse` and
-/// chime:`CreateAttendeeResponse` response.
+/// chime:`CreateAttendeeResponse` response and optional custom `URLRewriter` that will
+/// rewrite urls given to new urls.
 @objcMembers public class MeetingSessionConfiguration: NSObject {
     /// The id of the meeting the session is joining.
     public let meetingId: String
@@ -21,13 +22,26 @@ import Foundation
     /// The URLs the session uses to reach the meeting service.
     public let urls: MeetingSessionURLs
 
-    public init(createMeetingResponse: CreateMeetingResponse, createAttendeeResponse: CreateAttendeeResponse) {
+    public let urlRewriter: URLRewriter
+
+    public convenience init(createMeetingResponse: CreateMeetingResponse,
+                            createAttendeeResponse: CreateAttendeeResponse) {
+        self.init(createMeetingResponse: createMeetingResponse,
+                  createAttendeeResponse: createAttendeeResponse,
+                  urlRewriter: URLRewriterUtils.defaultUrlRewriter)
+    }
+
+    public init(createMeetingResponse: CreateMeetingResponse,
+                createAttendeeResponse: CreateAttendeeResponse,
+                urlRewriter: @escaping URLRewriter) {
         self.meetingId = createMeetingResponse.meeting.meetingId
         self.credentials = MeetingSessionCredentials(attendeeId: createAttendeeResponse.attendee.attendeeId,
                                                      joinToken: createAttendeeResponse.attendee.joinToken)
         self.urls = MeetingSessionURLs(audioFallbackUrl: createMeetingResponse.meeting.mediaPlacement.audioFallbackUrl,
                                        audioHostUrl: createMeetingResponse.meeting.mediaPlacement.audioHostUrl,
                                        turnControlUrl: createMeetingResponse.meeting.mediaPlacement.turnControlUrl,
-                                       signalingUrl: createMeetingResponse.meeting.mediaPlacement.signalingUrl)
+                                       signalingUrl: createMeetingResponse.meeting.mediaPlacement.signalingUrl,
+                                       urlRewriter: urlRewriter)
+        self.urlRewriter = urlRewriter
     }
 }